// src/application/dtos/auth/profile.dto.ts
import z from "zod";

export const profileResponseSchema = z.object({
   id: z.string(),
   name: z.string().optional(),
   surname: z.string().optional(),
   email: z.string().optional(),
   isMissingOnboarding: z.boolean(),
   individualAccount: z.object({
      id: z.string(),
      name: z.string().optional(),
      email: z.string(),
      phone: z.string().optional(),
      cpf: z.string().optional(),
      birthday: z.date().optional(),
      cep: z.string().optional(),
      state: z.string().optional(),
      address: z.string().optional(),
      neighborhood: z.string().optional(),
      city: z.string().optional(),
      number: z.string().optional(),
      createdAt: z.date().optional(),
      updatedAt: z.date().optional(),
   }).optional(),
   corporateAccount: z.object({
      id: z.string(),
      ownerName: z.string().optional(),
      socialReason: z.string().optional(),
      email: z.string().optional(),
      phone: z.string().optional(),
      cnpj: z.string().optional(),
      cep: z.string().optional(),
      state: z.string().optional(),
      address: z.string().optional(),
      neighborhood: z.string().optional(),
      city: z.string().optional(),
      number: z.string().optional(),
      createdAt: z.date().optional(),
      updatedAt: z.date().optional(),
   }).optional(),
   activeSubscription: z.object({
      id: z.string(),
      status: z.string(),
      startDate: z.date().optional(),
      endDate: z.date().optional(),
      nextPaymentDate: z.date().optional(),
      plan: z.object({
         id: z.string(),
         name: z.string(),
         description: z.string().optional(),
         amount: z.number(),
         periodicity: z.string(),
         isActive: z.boolean(),
      }),
      createdAt: z.date(),
      updatedAt: z.date(),
   }).optional(),
   createdAt: z.date().optional(),
   updatedAt: z.date().optional(),
});

export type ProfileResponseDto = z.infer<typeof profileResponseSchema>;